Details on handling the hpc/htc divide
Added details on handling the hpc/htc divide and possible natural consequences of not understanding or respecting it.
(Feel free to incorporate part, all or none of this. This "pull request" is really more of a suggestion. I'm still trying to figure out how to make this all work easiest for everyone without causing any undue surprises. I worry that running out of one's own personal group's space would be a common surprising situation.)