Can now extract gene name from gene id
Some gtf sources do not have a gene id field but instead have the gene
symbol inside the gene id in the format 'gene-[name]'. This commit
adds a script to check for this format and fix it so the gene name is
where cellranger expects it to be, if necessary.